I made work on the CSS and HTML of most pages, if not all.
There is lots of loose work, which will need to be fixed after the due date. Althought we will have done the main fucntion of the site done, it will have to be edited after the fact. CSS AND HTML work has been done to clean up the site, and will continue to do so.